refactor(gui): extract InputView component into input package
Deconstruct InputView into smaller, more modular widgets (toolbar,
header, list, rows, etc.) in a dedicated input/ subpackage to
improve readability and maintainability.
