+++ src/amplifyp/gui/views/input/__init__.py
@@ -1,0 +1,13 @@
+# Copyright (C) 2026 Fufu Fang
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+
+"""Package containing the DNA Template and Primer input components."""
+
+from .input_view import InputView
+
+__all__ = ["InputView"]
