+++ src/amplifyp/gui/views/input/primer_input.py
@@ -1,0 +1,461 @@
+# Copyright (C) 2026 Fufu Fang
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+
+"""Input component for DNA primers list and details."""
+
+from typing import Any
+import flet as ft
+
+from amplifyp.gui.settings import GUIColors, GUISettings
+from amplifyp.gui.user_data import GUIInput
+from amplifyp.gui.util import clean_sequence
+from .primer_row import PrimerRow
+from .primer_info_panel import PrimerInfoPanel
+from .primer_header import PrimerHeader
+from .primer_toolbar import PrimerToolbar
+from .primer_file_manager import PrimerFileManager
+from .primer_list import PrimerList
+
+
+
+class PrimerInput(ft.Container):  # type: ignore[misc]
+    """Input component for DNA primers list and details."""
+
+    def __init__(
+        self,
+        page: ft.Page,
+        settings: GUISettings,
+        input_data: GUIInput,
+        on_change_handler: Any,
+        handle_field_focus: Any,
+        handle_field_blur: Any,
+        handle_field_submit: Any,
+        clear_primers_callback: Any,
+    ) -> None:
+        """Initialize the PrimerInput component."""
+        super().__init__(expand=5)
+        self.app_page = page
+        self.settings = settings
+        self.input_data = input_data
+        self.on_change_handler = on_change_handler
+        self.handle_field_focus = handle_field_focus
+        self.handle_field_blur = handle_field_blur
+        self.handle_field_submit = handle_field_submit
+        self.clear_primers_callback = clear_primers_callback
+
+        self.focused_primer_index: int | None = None
+        self.validation_errors: list[str | None] = []
+
+        font_family = self.settings.get("font_family", "Roboto Mono")
+        self.name_column_width = 150.0
+
+        # Primer List Component
+        self.primers_list = PrimerList(self)
+
+        # Primer Header Component
+        self.primer_header = PrimerHeader(
+            settings=self.settings,
+            on_toggle_all=self._on_toggle_all_primers,
+            on_divider_pan=self._on_primer_divider_pan,
+            name_column_width=self.name_column_width,
+        )
+        # Compatibility links
+        self.all_primers_checkbox = self.primer_header.all_primers_checkbox
+        self.primers_header = self.primer_header.header_row
+        self.primers_header_container = self.primer_header
+
+        # File Manager Component
+        self.file_manager = PrimerFileManager(
+            page=self.app_page,
+            input_data=self.input_data,
+            on_update_ui=self.update_ui,
+            on_change_handler=self.on_change_handler,
+            show_snackbar=self._show_snackbar,
+        )
+
+        # Primer Toolbar Component
+        self.primer_toolbar = PrimerToolbar(
+            on_save=self._save_primers_click,
+            on_load=self._load_primers_click,
+            on_clear=self.clear_primers_callback,
+        )
+        # Compatibility links
+        self.clear_primers_button = self.primer_toolbar.clear_button
+        self.save_primers_button = self.primer_toolbar.save_button
+        self.load_primers_button = self.primer_toolbar.load_button
+
+        # Primer Info Panel
+        self.primer_info_panel = PrimerInfoPanel(
+            settings=self.settings, font_family=font_family
+        )
+
+        self.content = ft.Column(
+            [
+                ft.Row(
+                    [
+                        ft.Text("Primers", weight=ft.FontWeight.BOLD),
+                        self.primer_toolbar,
+                    ],
+                    alignment=ft.MainAxisAlignment.SPACE_BETWEEN,
+                    height=40,
+                ),
+                ft.Container(
+                    content=ft.Column(
+                        [
+                            self.primers_header_container,
+                            self.primers_list,
+                        ],
+                        expand=True,
+                        spacing=0,
+                    ),
+                    expand=True,
+                    border=ft.Border.all(1, GUIColors.OUTLINE),
+                    border_radius=5,
+                    padding=0,
+                ),
+                self.primer_info_panel,
+            ],
+            expand=True,
+            spacing=5,
+        )
+
+    # Delegate properties of the info panel for test and backward compatibility
+    @property
+    def info_header(self) -> ft.Container:
+        """Get the primer info header container."""
+        return self.primer_info_panel.info_header
+
+    @property
+    def info_seq_text(self) -> ft.Text:
+        """Get the primer info sequence text control."""
+        return self.primer_info_panel.info_seq_text
+
+    @property
+    def info_tm_text(self) -> ft.Text:
+        """Get the primer info Tm text control."""
+        return self.primer_info_panel.info_tm_text
+
+    @property
+    def info_pairs_text(self) -> ft.Text:
+        """Get the primer info AT/GC pairs text control."""
+        return self.primer_info_panel.info_pairs_text
+
+    @property
+    def info_redundancy_text(self) -> ft.Text:
+        """Get the primer info redundancy text control."""
+        return self.primer_info_panel.info_redundancy_text
+
+    @property
+    def info_dimer_text(self) -> ft.Text:
+        """Get the primer info dimer text control."""
+        return self.primer_info_panel.info_dimer_text
+
+    def _handle_row_click(self, idx: int, name_edit: ft.TextField) -> None:
+        """Handle clicking on the row container.
+
+        Selects the row and focuses the name field.
+        """
+        self.focused_primer_index = idx
+        self._update_row_highlights()
+        self._update_primer_info_panel()
+        name_edit.focus()
+
+    def _move_primer(self, idx: int, direction: int) -> None:
+        """Move primer at idx up (direction=-1) or down (direction=1)."""
+        parent_view = getattr(self.on_change_handler, "__self__", None)
+        if parent_view is not None:
+            if getattr(parent_view, "_focus_timer", None) is not None:
+                parent_view._focus_timer.cancel()
+                parent_view._focus_timer = None
+            parent_view._skip_blur_timer = True
+
+        self.sync_to_state(rebuild_if_needed=False)
+        primers = self.input_data.primers
+        target_idx = idx + direction
+
+        # Swap only if both indices are valid filled primers.
+        if 0 <= idx < len(primers) and 0 <= target_idx < len(primers):
+            primers[idx], primers[target_idx] = (
+                primers[target_idx],
+                primers[idx],
+            )
+            if self.focused_primer_index == idx:
+                self.focused_primer_index = target_idx
+            elif self.focused_primer_index == target_idx:
+                self.focused_primer_index = idx
+
+            self.update_ui()
+            if self.on_change_handler:
+                self.on_change_handler(None)
+
+    def _extract_primer_data_from_ui(self) -> list[dict[str, Any]]:
+        """Extract primer data from UI controls."""
+        ui_primers = []
+        for row in self.primers_list.controls:
+            if not isinstance(row, PrimerRow):
+                continue
+
+            name_val = str(row.name_field.value or "").strip()
+            seq_val = clean_sequence(str(row.seq_field.value or ""))
+            is_active = bool(row.checkbox.value)
+
+            ui_primers.append(
+                {
+                    "name": name_val,
+                    "seq": seq_val,
+                    "active": is_active,
+                    "container": row,
+                    "checkbox": row.checkbox,
+                }
+            )
+        return ui_primers
+
+    def sync_to_state(self, rebuild_if_needed: bool = True) -> bool:
+        """Sync current UI controls back to the central state.
+
+        Returns True if a UI rebuild is needed.
+        """
+        ui_primers = self._extract_primer_data_from_ui()
+        filtered_ui_primers, should_rebuild = self._apply_activation_rules(
+            ui_primers
+        )
+
+        dup_indices = self._get_duplicate_indices_for_list(filtered_ui_primers)
+        primers = []
+        for p in filtered_ui_primers:
+            container = p["container"]
+            c_idx = container.data
+            is_dup = c_idx in dup_indices
+            new_color = GUIColors.DUPLICATE_BG if is_dup else None
+            if container.bgcolor != new_color:
+                container.bgcolor = new_color
+
+            primers.append(
+                {
+                    "name": p["name"],
+                    "seq": p["seq"],
+                    "active": p["active"],
+                }
+            )
+
+        # Run background primer construction/validation
+        new_validation_errors = []
+        for p in primers:
+            new_validation_errors.append(
+                PrimerRow.validate(p["name"], p["seq"])
+            )
+
+        if getattr(self, "validation_errors", []) != new_validation_errors:
+            should_rebuild = True
+
+        # If the number of non-empty primers changed, we rebuild
+        if len(primers) != len(
+            [
+                p
+                for p in self.input_data.primers
+                if p.get("name", "").strip() or p.get("seq", "").strip()
+            ]
+        ):
+            should_rebuild = True
+
+        self.input_data.primers = primers
+        self.validation_errors = new_validation_errors
+        if should_rebuild and rebuild_if_needed:
+            self.update_ui()
+        else:
+            self._update_header_checkbox_state()
+
+        return should_rebuild
+
+    def _apply_activation_rules(
+        self,
+        ui_primers: list[dict[str, Any]],
+    ) -> tuple[list[dict[str, Any]], bool]:
+        """Apply auto-activation, auto-inactivation and deletion rules."""
+        from amplifyp.dna import Primer
+        prev_primers = self.input_data.primers
+        filtered_primers = []
+        should_rebuild = False
+
+        for i, p in enumerate(ui_primers):
+            name_val = p["name"]
+            seq_val = p["seq"]
+            is_active = p["active"]
+            checkbox = p.get("checkbox")
+
+            # Deletion rule: if both are empty, delete it
+            if not name_val and not seq_val:
+                should_rebuild = True
+                continue
+
+            prev_primer = prev_primers[i] if i < len(prev_primers) else None
+            prev_name = (
+                str(prev_primer.get("name", "")).strip() if prev_primer else ""
+            )
+            prev_seq = (
+                clean_sequence(str(prev_primer.get("seq", "")))
+                if prev_primer
+                else ""
+            )
+
+            prev_seq_invalid = False
+            if prev_seq:
+                try:
+                    Primer(sequence=prev_seq, name=prev_name)
+                except ValueError:
+                    prev_seq_invalid = True
+
+            # Auto-activation rule
+            if (not prev_name or not prev_seq or prev_seq_invalid) and (
+                name_val and seq_val
+            ):
+                is_valid = True
+                try:
+                    Primer(sequence=seq_val, name=name_val)
+                except ValueError:
+                    is_valid = False
+                if is_valid:
+                    is_active = True
+                    if checkbox is not None:
+                        checkbox.value = True
+                    should_rebuild = True
+
+            # Auto-inactivation rule
+            if not name_val or not seq_val:
+                if is_active:
+                    is_active = False
+                    if checkbox is not None:
+                        checkbox.value = False
+                    should_rebuild = True
+
+            p["active"] = is_active
+            filtered_primers.append(p)
+
+        return filtered_primers, should_rebuild
+
+    def update_ui(self) -> None:
+        """Update Flet UI controls to match the central state."""
+        self.primers_list.update_list_ui()
+
+    def _on_primer_divider_pan(self, e: ft.DragUpdateEvent) -> None:
+        """Handle dragging the vertical divider between Name and Sequence."""
+        delta_x = getattr(e.local_delta, "x", 0.0) if e.local_delta else 0.0
+        self.name_column_width = max(
+            80.0, min(300.0, self.name_column_width + delta_x)
+        )
+        # Update the width of the Name header control
+        self.primers_header.controls[2].width = self.name_column_width
+        # Update the width of all Name TextFields in the list controls
+        for row in self.primers_list.controls:
+            if isinstance(row, PrimerRow):
+                row.name_field.width = self.name_column_width
+        self.app_page.update()
+
+    def _on_toggle_all_primers(self, e: Any) -> None:
+        """Toggle all primers active/inactive based on tri-state checkbox."""
+        primers = self.input_data.primers
+        if not primers:
+            return
+        non_empty = [
+            p
+            for p in primers
+            if str(p.get("name", "")).strip()
+            or clean_sequence(str(p.get("seq", ""))).strip()
+        ]
+        if not non_empty:
+            return
+        all_active = all(p.get("active", True) for p in non_empty)
+        target_active = not all_active
+        for p in non_empty:
+            p["active"] = target_active
+        self.update_ui()
+        if self.on_change_handler:
+            self.on_change_handler(e)
+
+    def _update_header_checkbox_state(self) -> None:
+        """Update the header checkbox to reflect the current primer states."""
+        primers = self.input_data.primers
+        non_empty = [
+            p
+            for p in primers
+            if str(p.get("name", "")).strip()
+            or clean_sequence(str(p.get("seq", ""))).strip()
+        ]
+        if not non_empty:
+            self.all_primers_checkbox.value = None
+        elif all(p.get("active", True) for p in non_empty):
+            self.all_primers_checkbox.value = True
+        elif all(not p.get("active", True) for p in non_empty):
+            self.all_primers_checkbox.value = False
+        else:
+            self.all_primers_checkbox.value = None
+        if self.app_page:
+            self.app_page.update()
+
+    def _get_duplicate_indices_for_list(
+        self, primers: list[dict[str, Any]]
+    ) -> set[int]:
+        """Find and return indices of duplicate primers by name/sequence."""
+        names_count: dict[str, int] = {}
+        seqs_count: dict[str, int] = {}
+        for p in primers:
+            n_lower = str(p.get("name", "")).strip().lower()
+            s_lower = clean_sequence(str(p.get("seq", ""))).lower()
+            if n_lower:
+                names_count[n_lower] = names_count.get(n_lower, 0) + 1
+            if s_lower:
+                seqs_count[s_lower] = seqs_count.get(s_lower, 0) + 1
+
+        dup_indices = set()
+        for idx, p in enumerate(primers):
+            n_lower = str(p.get("name", "")).strip().lower()
+            s_lower = clean_sequence(str(p.get("seq", ""))).lower()
+            if (n_lower and names_count.get(n_lower, 0) > 1) or (
+                s_lower and seqs_count.get(s_lower, 0) > 1
+            ):
+                # If the item has a container reference, use container.data,
+                # otherwise use idx
+                c = p.get("container")
+                c_idx = c.data if (c is not None and hasattr(c, "data")) else idx
+                dup_indices.add(c_idx)
+        return dup_indices
+
+    def _get_duplicate_indices(self) -> set[int]:
+        """Find indices of primers with duplicate names or sequences."""
+        return self._get_duplicate_indices_for_list(self.input_data.primers)
+
+    def _update_row_highlights(self) -> None:
+        """Update background colors of all row containers.
+
+        Highlights rows based on selection and duplicates.
+        """
+        self.primers_list.update_row_highlights()
+
+    def _update_primer_info_panel(self) -> None:
+        """Update the primer information panel based on the focused primer."""
+        self.primer_info_panel.update_panel(
+            focused_idx=self.focused_primer_index,
+            input_data=self.input_data,
+            app_page=self.app_page,
+            on_update_highlights=self._update_row_highlights,
+        )
+
+    async def _load_primers_click(self, e: ft.ControlEvent) -> None:
+        """Open file picker to load primers from CSV/TSV file."""
+        await self.file_manager.load_primers_click(e)
+
+    async def _save_primers_click(self, e: ft.ControlEvent) -> None:
+        """Save active primers to a CSV file."""
+        await self.file_manager.save_primers_click(e)
+
+    def _show_snackbar(self, message: str) -> None:
+        """Show a snackbar message."""
+        self.app_page.overlay.append(ft.SnackBar(ft.Text(message), open=True))
+        self.app_page.update()
+
+
