+++ src/amplifyp/gui/views/input/primer_row.py
@@ -1,0 +1,197 @@
+# Copyright (C) 2026 Fufu Fang
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+
+"""A single row representing a primer in the list."""
+
+from typing import Any
+import flet as ft
+
+from amplifyp.gui.settings import GUIColors
+
+
+class PrimerRow(ft.Container):  # type: ignore[misc]
+    """A single row representing a primer in the list."""
+
+    def __init__(
+        self,
+        idx: int,
+        name: str,
+        seq: str,
+        is_active: bool,
+        is_dup: bool,
+        error_message: str | None,
+        font_family: str,
+        name_column_width: float,
+        on_change_handler: Any,
+        handle_field_focus: Any,
+        handle_field_blur: Any,
+        handle_field_submit: Any,
+        on_row_click: Any,
+        on_move_primer: Any,
+        on_divider_pan: Any,
+        is_focused: bool,
+        is_last_row: bool,
+        is_penultimate_row: bool,
+    ) -> None:
+        super().__init__(
+            data=idx,
+            bgcolor=GUIColors.DUPLICATE_BG if is_dup else None,
+            padding=0,
+            height=30 if not error_message else None,
+        )
+        self.idx = idx
+        self.checkbox = ft.Checkbox(
+            value=is_active,
+            on_change=on_change_handler,
+        )
+        self.checkbox_container = ft.Container(
+            content=self.checkbox,
+            width=55,
+            height=30,
+            alignment=ft.Alignment(0, 0),
+        )
+        self.name_field = ft.TextField(
+            value=name,
+            hint_text="New Primer Name",
+            dense=True,
+            content_padding=ft.Padding(5, 0, 0, 0),
+            height=30,
+            width=name_column_width,
+            border=ft.InputBorder.NONE,
+            data=idx,
+            on_focus=handle_field_focus,
+            on_blur=handle_field_blur,
+            on_submit=handle_field_submit,
+        )
+        self.seq_field = ft.TextField(
+            value=seq,
+            hint_text="New Primer Sequence",
+            dense=True,
+            content_padding=ft.Padding(
+                5,
+                0,
+                60 if is_focused and not is_last_row else 0,
+                0,
+            ),
+            height=30 if not error_message else None,
+            border=ft.InputBorder.NONE,
+            text_style=ft.TextStyle(font_family=font_family),
+            data=idx,
+            on_focus=handle_field_focus,
+            on_blur=handle_field_blur,
+            on_submit=handle_field_submit,
+        )
+        if error_message:
+            self.seq_field.error = error_message
+
+        self.divider = ft.GestureDetector(
+            on_pan_update=on_divider_pan,
+            content=ft.Container(
+                width=4,
+                bgcolor=GUIColors.DIVIDER_GREY,
+                margin=0,
+                height=30,
+            ),
+            mouse_cursor=ft.MouseCursor.RESIZE_LEFT_RIGHT,
+        )
+
+        self.active_divider = ft.Container(
+            width=4,
+            bgcolor=GUIColors.DIVIDER_GREY,
+            margin=0,
+            height=30,
+        )
+
+        self.seq_stack = ft.Stack(
+            [
+                self.seq_field,
+            ],
+            expand=True,
+        )
+
+        self.reorder_container = None
+        if not is_last_row:
+            up_button = ft.IconButton(
+                icon=ft.Icons.ARROW_UPWARD,
+                icon_size=16,
+                width=24,
+                height=24,
+                padding=0,
+                tooltip="Move Up",
+                disabled=(idx == 0),
+                on_click=lambda e: on_move_primer(idx, -1),
+            )
+            down_button = ft.IconButton(
+                icon=ft.Icons.ARROW_DOWNWARD,
+                icon_size=16,
+                width=24,
+                height=24,
+                padding=0,
+                tooltip="Move Down",
+                disabled=is_penultimate_row,
+                on_click=lambda e: on_move_primer(idx, 1),
+            )
+            self.reorder_controls = ft.Row(
+                [up_button, down_button],
+                spacing=2,
+                alignment=ft.MainAxisAlignment.CENTER,
+            )
+            self.reorder_container = ft.Container(
+                content=self.reorder_controls,
+                right=10,
+                top=3,
+                visible=is_focused,
+            )
+            self.seq_stack.controls.append(self.reorder_container)
+
+        self.content = ft.Row(
+            [
+                self.checkbox_container,
+                self.active_divider,
+                self.name_field,
+                self.divider,
+                self.seq_stack,
+            ],
+            spacing=0,
+            vertical_alignment=ft.CrossAxisAlignment.START,
+        )
+        self.on_click = lambda e: on_row_click(idx, self.name_field)
+
+    def update_highlight_and_reorder(self, is_focused: bool, is_dup: bool) -> None:
+        """Update the background color and reorder buttons layout."""
+        if is_focused:
+            self.bgcolor = GUIColors.SELECTED_ROW_BG
+        elif is_dup:
+            self.bgcolor = GUIColors.DUPLICATE_BG
+        else:
+            self.bgcolor = None
+
+        if self.reorder_container is not None:
+            self.reorder_container.visible = is_focused
+            self.seq_field.content_padding = ft.Padding(
+                5, 0, 60 if is_focused else 0, 0
+            )
+
+    def set_error(self, err: str | None) -> None:
+        """Set or clear the error message on the sequence field and adjust height."""
+        self.seq_field.error = err
+        self.seq_field.height = 30 if not err else None
+        self.height = 30 if not err else None
+
+    @staticmethod
+    def validate(name: str, seq: str) -> str | None:
+        """Validate a primer sequence and name, returning an error message if invalid."""
+        if not seq:
+            return None
+        try:
+            from amplifyp.dna import Primer
+            Primer(sequence=seq, name=name)
+            return None
+        except ValueError as ex:
+            return str(ex)
+
